refactor: useSignature custom hook 생성
